Store decoding errors in individual paywall
